requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.4.0
astunparse==1.6.3
bcrypt==4.0.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
cli_helpers==2.3.1
click==8.1.7
configobj==5.0.8
contourpy==1.3.0
cycler==0.12.1
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.9.0
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.66.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
keras==3.5.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
mtcnn==0.1.1
namex==0.0.8
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.3
packaging==24.1
pandas==2.2.2
passlib==1.7.4
pgcli==4.1.0
pgspecial==2.1.2
pillow==10.4.0
prompt_toolkit==3.0.47
protobuf==4.25.5
psutil==6.0.0
psycopg==3.1.19
psycopg-binary==3.1.19
py-cpuinfo==9.0.0
pyasn1==0.6.1
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.4
pypng==0.20220715.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
requests==2.32.3
retina-face==0.0.17
rich==13.7.1
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.31
sqlparse==0.5.0
starlette==0.37.2
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
tf_keras==2.17.0
threadpoolctl==3.5.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
typer==0.12.3
typing_extensions==4.12.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.4
wrapt==1.16.0
